Basic matrix & vector methods

Matrix functions

GraphBLASInterface.GrB_Matrix_new — Function.

GrB_Matrix_new(A, type, nrows, ncols)

Initialize a matrix with specified domain and dimensions.

source

GraphBLASInterface.GrB_Matrix_build — Function.

GrB_Matrix_build(C, I, J, X, nvals, dup)

Store elements from tuples into a matrix.

source

GraphBLASInterface.GrB_Matrix_nrows — Function.

GrB_Matrix_nrows(A)

Return the number of rows in a matrix if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Matrix_ncols — Function.

GrB_Matrix_ncols(A)

Return the number of columns in a matrix if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Matrix_nvals — Function.

GrB_Matrix_nvals(A)

Return the number of stored elements in a matrix if successful. Else return GrB_Info error code..

source

GraphBLASInterface.GrB_Matrix_setElement — Function.

GrB_Matrix_setElement(C, X, I, J)

Set one element of a matrix to a given value, C[I][J] = X.

source

GraphBLASInterface.GrB_Matrix_extractElement — Function.

GrB_Matrix_extractElement(A, row_index, col_index)

Return element of a matrix at a given index (A[rowindex][colindex]) if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Matrix_extractTuples — Function.

GrB_Matrix_extractTuples(A)

Return tuples stored in a matrix if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Matrix_dup — Function.

GrB_Matrix_dup(C, A)

Initialize a new matrix with the same domain, dimensions, and contents as another matrix.

source

GraphBLASInterface.GrB_Matrix_clear — Function.

GrB_Matrix_clear(A)

Remove all elements from a matrix.

source

Vector functions

GraphBLASInterface.GrB_Vector_new — Function.

GrB_Vector_new(v, type, n)

Initialize a vector with specified domain and size.

source

GraphBLASInterface.GrB_Vector_build — Function.

GrB_Vector_build(w, I, X, nvals, dup)

Store elements from tuples into a vector.

source

GraphBLASInterface.GrB_Vector_size — Function.

GrB_Vector_size(v)

Return the size of a vector if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Vector_nvals — Function.

GrB_Vector_nvals(v)

Return the number of stored elements in a vector if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Vector_setElement — Function.

GrB_Vector_setElement(w, x, i)

Set one element of a vector to a given value, w[i] = x.

source

GraphBLASInterface.GrB_Vector_extractElement — Function.

GrB_Vector_extractElement(v, i)

Return element of a vector at a given index (v[i]) if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Vector_extractTuples — Function.

GrB_Vector_extractTuples(v)

Return tuples stored in a vector if successful. Else return GrB_Info error code.

source

GraphBLASInterface.GrB_Vector_dup — Function.

GrB_Vector_dup(w, u)

Initialize a vector with the same domain, size, and contents as another vector.

source

GraphBLASInterface.GrB_Vector_clear — Function.

GrB_Vector_clear(v)

Remove all the elements (tuples) from a vector.

source