Operations

GraphBLAS operations

Multiplication

GraphBLASInterface.GrB_mxm — Function.

GrB_mxm(C, Mask, accum, semiring, A, B, desc)

Multiplies a matrix with another matrix on a semiring. The result is a matrix.

source

GraphBLASInterface.GrB_vxm — Function.

GrB_vxm(w, mask, accum, semiring, u, A, desc)

Multiplies a (row)vector with a matrix on an semiring. The result is a vector.

source

GraphBLASInterface.GrB_mxv — Function.

GrB_mxv(w, mask, accum, semiring, A, u, desc)

Multiplies a matrix by a vector on a semiring. The result is a vector.

source

Element-wise multiplication

GraphBLASInterface.GrB_eWiseMult — Function.

GrB_eWiseMult(C, mask, accum, op, A, B, desc)

Generic method for element-wise matrix and vector operations: using set intersection.

GrB_eWiseMult computes C<Mask> = accum (C, A .* B), where pairs of elements in two matrices (or vectors) are pairwise "multiplied" with C(i, j) = mult (A(i, j), B(i, j)). The "multiplication" operator can be any binary operator. The pattern of the result T = A .* B is the set intersection (not union) of A and B. Entries outside of the intersection are not computed. This is primary difference with GrB_eWiseAdd. The input matrices A and/or B may be transposed first, via the descriptor. For a semiring, the mult operator is the semiring's multiply operator; this differs from the eWiseAdd methods which use the semiring's add operator instead.

source

GraphBLASInterface.GrB_eWiseMult_Vector_Semiring — Function.

GrB_eWiseMult_Vector_Semiring(w, mask, accum, semiring, u, v, desc)

Compute element-wise vector multiplication using semiring. Semiring's multiply operator is used. w<mask> = accum (w, u .* v)

source

GraphBLASInterface.GrB_eWiseMult_Vector_Monoid — Function.

GrB_eWiseMult_Vector_Monoid(w, mask, accum, monoid, u, v, desc)

Compute element-wise vector multiplication using monoid. w<mask> = accum (w, u .* v)

source

GraphBLASInterface.GrB_eWiseMult_Vector_BinaryOp — Function.

GrB_eWiseMult_Vector_BinaryOp(w, mask, accum, mult, u, v, desc)

Compute element-wise vector multiplication using binary operator. w<mask> = accum (w, u .* v)

source

GraphBLASInterface.GrB_eWiseMult_Matrix_Semiring — Function.

GrB_eWiseMult_Matrix_Semiring(C, Mask, accum, semiring, A, B, desc)

Compute element-wise matrix multiplication using semiring. Semiring's multiply operator is used. C<Mask> = accum (C, A .* B)

source

GraphBLASInterface.GrB_eWiseMult_Matrix_Monoid — Function.

GrB_eWiseMult_Matrix_Monoid(C, Mask, accum, monoid, A, B, desc)

Compute element-wise matrix multiplication using monoid. C<Mask> = accum (C, A .* B)

source

GraphBLASInterface.GrB_eWiseMult_Matrix_BinaryOp — Function.

GrB_eWiseMult_Matrix_BinaryOp(C, Mask, accum, mult, A, B, desc)

Compute element-wise matrix multiplication using binary operator. C<Mask> = accum (C, A .* B)

source

Element-wise addition

GraphBLASInterface.GrB_eWiseAdd — Function.

GrB_eWiseAdd(C, mask, accum, op, A, B, desc)

Generic method for element-wise matrix and vector operations: using set union.

GrB_eWiseAdd computes C<Mask> = accum (C, A + B), where pairs of elements in two matrices (or two vectors) are pairwise "added". The "add" operator can be any binary operator. With the plus operator, this is the same matrix addition in conventional linear algebra. The pattern of the result T = A + B is the set union of A and B. Entries outside of the union are not computed. That is, if both A(i, j) and B(i, j) are present in the pattern of A and B, then T(i, j) = A(i, j) "+" B(i, j). If only A(i, j) is present then T(i, j) = A (i, j) and the "+" operator is not used. Likewise, if only B(i, j) is in the pattern of B but A(i, j) is not in the pattern of A, then T(i, j) = B(i, j). For a semiring, the mult operator is the semiring's add operator.

source

GraphBLASInterface.GrB_eWiseAdd_Vector_Semiring — Function.

GrB_eWiseAdd_Vector_Semiring(w, mask, accum, semiring, u, v, desc)

Compute element-wise vector addition using semiring. Semiring's add operator is used. w<mask> = accum (w, u + v)

source

GraphBLASInterface.GrB_eWiseAdd_Vector_Monoid — Function.

GrB_eWiseAdd_Vector_Monoid(w, mask, accum, monoid, u, v, desc)

Compute element-wise vector addition using monoid. w<mask> = accum (w, u + v)

source

GraphBLASInterface.GrB_eWiseAdd_Vector_BinaryOp — Function.

GrB_eWiseAdd_Vector_BinaryOp(w, mask, accum, add, u, v, desc)

Compute element-wise vector addition using binary operator. w<mask> = accum (w, u + v)

source

GraphBLASInterface.GrB_eWiseAdd_Matrix_Semiring — Function.

GrB_eWiseAdd_Matrix_Semiring(C, Mask, accum, semiring, A, B, desc)

Compute element-wise matrix addition using semiring. Semiring's add operator is used. C<Mask> = accum (C, A + B)

source

GraphBLASInterface.GrB_eWiseAdd_Matrix_Monoid — Function.

GrB_eWiseAdd_Matrix_Monoid(C, Mask, accum, monoid, A, B, desc)

Compute element-wise matrix addition using monoid. C<Mask> = accum (C, A + B)

source

GraphBLASInterface.GrB_eWiseAdd_Matrix_BinaryOp — Function.

GrB_eWiseAdd_Matrix_BinaryOp(C, Mask, accum, add, A, B, desc)

Compute element-wise matrix addition using binary operator. C<Mask> = accum (C, A + B)

source

Extract

GraphBLASInterface.GrB_extract — Function.

GrB_extract(arg1, Mask, accum, arg4, ...)

Generic matrix/vector extraction.

source

GraphBLASInterface.GrB_Vector_extract — Function.

GrB_Vector_extract(w, mask, accum, u, I, ni, desc)

Extract a sub-vector from a larger vector as specified by a set of indices. The result is a vector whose size is equal to the number of indices.

source

GraphBLASInterface.GrB_Matrix_extract — Function.

GrB_Matrix_extract(C, Mask, accum, A, I, ni, J, nj, desc)

Extract a sub-matrix from a larger matrix as specified by a set of row indices and a set of column indices. The result is a matrix whose size is equal to size of the sets of indices.

source

GraphBLASInterface.GrB_Col_extract — Function.

GrB_Col_extract(w, mask, accum, A, I, ni, j, desc)

Extract from one column of a matrix into a vector. With the transpose descriptor for the source matrix, elements of an arbitrary row of the matrix can be extracted with this function as well.

source

Assign

GraphBLASInterface.GrB_assign — Function.

GrB_assign(arg1, Mask, accum, arg4, arg5, ...)

Generic method for submatrix/subvector assignment.

source

GraphBLASInterface.GrB_Vector_assign — Function.

GrB_Vector_assign(w, mask, accum, u, I, ni, desc)

Assign values from one GraphBLAS vector to a subset of a vector as specified by a set of indices. The size of the input vector is the same size as the index array provided.

source

GrB_Vector_assign(w, mask, accum, x, I, ni, desc)

Assign the same value to a specified subset of vector elements. With the use of GrB_ALL, the entire destination vector can be filled with the constant. ```

source

GraphBLASInterface.GrB_Matrix_assign — Function.

GrB_Matrix_assign(C, Mask, accum, A, I, ni, J, nj, desc)

Assign values from one GraphBLAS matrix to a subset of a matrix as specified by a set of indices. The dimensions of the input matrix are the same size as the row and column index arrays provided.

source

GrB_Matrix_assign(C, Mask, accum, x, I, ni, J, nj, desc)

Assign the same value to a specified subset of matrix elements. With the use of GrB_ALL, the entire destination matrix can be filled with the constant.

source

GraphBLASInterface.GrB_Col_assign — Function.

GrB_Col_assign(C, Mask, accum, u, I, ni, j, desc)

Assign the contents of a vector to a subset of elements in one column of a matrix. Note that since the output cannot be transposed, a different variant of assign is provided to assign to a row of matrix.

source

GraphBLASInterface.GrB_Row_assign — Function.

GrB_Row_assign(C, mask, accum, u, i, J, nj, desc)

Assign the contents of a vector to a subset of elements in one row of a matrix. Note that since the output cannot be transposed, a different variant of assign is provided to assign to a column of a matrix.

source

Apply

GraphBLASInterface.GrB_apply — Function.

GrB_apply(C, Mask, accum, op, A, desc)

Generic matrix/vector apply.

source

GraphBLASInterface.GrB_Vector_apply — Function.

GrB_Vector_apply(w, mask, accum, op, u, desc)

Compute the transformation of the values of the elements of a vector using a unary function.

source

GraphBLASInterface.GrB_Matrix_apply — Function.

GrB_Matrix_apply(C, Mask, accum, op, A, desc)

Compute the transformation of the values of the elements of a matrix using a unary function.

source

Reduce

GraphBLASInterface.GrB_reduce — Function.

GrB_reduce(arg1, arg2, arg3, arg4, ...)

Generic method for matrix/vector reduction to a vector or scalar.

source

GraphBLASInterface.GrB_Matrix_reduce_Monoid — Function.

GrB_Matrix_reduce_Monoid(w, mask, accum, monoid, A, desc)

Reduce the entries in a matrix to a vector. By default these methods compute a column vector w such that w(i) = sum(A(i,:)), where "sum" is a commutative and associative monoid with an identity value. A can be transposed, which reduces down the columns instead of the rows.

source

GraphBLASInterface.GrB_Matrix_reduce_BinaryOp — Function.

GrB_Matrix_reduce_BinaryOp(w, mask, accum, op, A, desc)

Reduce the entries in a matrix to a vector. By default these methods compute a column vector w such that w(i) = sum(A(i,:)), where "sum" is a commutative and associative binary operator. A can be transposed, which reduces down the columns instead of the rows.

source

GraphBLASInterface.GrB_Vector_reduce — Function.

GrB_Vector_reduce(monoid, u, desc)

Reduce entries in a vector to a scalar. All entries in the vector are "summed" using the reduce monoid, which must be associative (otherwise the results are undefined). If the vector has no entries, the result is the identity value of the monoid.

source

GraphBLASInterface.GrB_Matrix_reduce — Function.

GrB_Matrix_reduce(monoid, A, desc)

Reduce entries in a matrix to a scalar. All entries in the matrix are "summed" using the reduce monoid, which must be associative (otherwise the results are undefined). If the matrix has no entries, the result is the identity value of the monoid.

source

Transpose

GraphBLASInterface.GrB_transpose — Function.

GrB_transpose(C, Mask, accum, A, desc)

Compute a new matrix that is the transpose of the source matrix.

source